Add remote code to interface

+++ src/dash/remote.c
@@ -2,6 +2,7 @@
 #include <unistd.h>
 #include <sys/wait.h>
 #include <string.h>
+#define _POSIX_SOURCE
 #include <signal.h>
 
 #define DEFAULT_ADDR "192.168.1.44"
@@ -9,11 +10,19 @@
 #define ADDR_BUFSIZE 32
 #define DEFAULT_EXE "/var/arcs/arcs"
 
-pid_t pid;
+pid_t pid = 0;
+
+char *addrstr() {
+	return r_args->r_addr == NULL ? DEFAULT_ADDR : r_args->r_addr;
+}
+
+char *userstr() {
+	return r_args->r_user == NULL ? DEFAULT_USER : r_args->r_user;
+}
 
 bool try_connect() {
-	char *addr = r_args->r_addr == NULL ? DEFAULT_ADDR : r_args->r_addr;
-	char *user = r_args->r_user == NULL ? DEFAULT_USER : r_args->r_user;
+	char *addr = addrstr();
+	char *user = userstr();
 
 	int statval;
 
@@ -32,12 +41,13 @@ bool try_connect() {
 		}
 #endif
 		snprintf(arg, ADDR_BUFSIZE, "%s@%s", user, addr);
-		char *args[] = {"ssh", arg, "exit",  0};
+		char *args[] = {"ssh", "-o", "ConnectTimeout=2", arg, "exit",  0};
 		char *env[] = {0};
 		execve("/usr/bin/ssh", args, env);
 	}
 	else {
 		waitpid(pid, &statval, 0);
+		pid = 0;
 		if(WIFEXITED(statval))
 			return (WEXITSTATUS(statval) == 0);
 		else
@@ -48,8 +58,8 @@ bool try_connect() {
 }
 
 void start_remote() {
-	char *addr = r_args->r_addr == NULL ? DEFAULT_ADDR : r_args->r_addr;
-	char *user = r_args->r_user == NULL ? DEFAULT_USER : r_args->r_user;
+	char *addr = addrstr();
+	char *user = userstr();
 	char *exec = r_args->r_exec == NULL ? DEFAULT_EXE : r_args->r_exec;
 
 	pid = fork();
@@ -74,5 +84,7 @@ void start_remote() {
 }
 
 void kill_remote() {
-	kill(pid, SIGKILL);
+	if(pid != 0)
+		kill(pid, SIGKILL);
+	pid = 0;
 }
